refactor(export): move TensorRT-LLM-only export code into export/trtllm
modelopt/torch/export mixed the framework-agnostic HF/Megatron export path
with code reachable only from export_tensorrt_llm_checkpoint. Split the two so
the TensorRT-LLM checkpoint format stays out of the shared path.

Moved whole, having no non-TRT-LLM importer: model_config_export,
model_config_utils, postprocess, distribute, tensorrt_llm_utils,
tensorrt_llm_type, hf_config_map, mcore_config_map.

Split where the module was mixed:

- model_config.py: the ModelConfig dataclasses and the LINEAR_*/LAYERNORM_*
  checkpoint-layout constants move to trtllm/model_config.py. The shared
  QUANTIZATION_*/KV_CACHE_*/FUSION_FREE_FORMATS constants move to a new leaf
  module, quant_format.py.
- layer_utils.py: the build_*_config builders and the enc/dec helpers move to
  trtllm/layer_utils.py. The module-shape predicates and MoE quantizer helpers
  (is_moe, is_quantlinear, get_experts_list, sync_moe_gate_up_amax, ...) stay.
- quant_utils.py: get_scaling_factor_from_weight and resmooth_and_get_scale,
  used only by trtllm/postprocess.py, move to trtllm/quant_utils.py.

trtllm/ reaches into the parent only through quant_format, quant_utils and
layer_utils, and nothing in the parent imports trtllm/, so the dependency is
one-way and the HF path can no longer drift back into checkpoint concerns.

Backward breaking: TensorRT-LLM export is no longer re-exported from
modelopt.torch.export. Import export_tensorrt_llm_checkpoint and
torch_to_tensorrt_llm_checkpoint from modelopt.torch.export.trtllm, and the
ModelConfig dataclasses from modelopt.torch.export.trtllm.model_config.

Testing: pre-commit passes on all changed files (ruff, ruff-format, mypy,
bandit, markdownlint). tests/unit/torch/export: 189 passed. Full
tests/unit/torch: 2367 passed with no export failures; the 45 failures are
pre-existing environment issues, a deepspeed circular import and a read-only
HF cache, confirmed by their error text. The GPU suites
(tests/gpu/torch/export, tests/gpu_trtllm) were not run, as no GPU is
available in this environment.
