+++ examples/deepseek/deepseek_v3/ptq.py
@@ -53,7 +53,7 @@ from tqdm import tqdm
 from transformers import AutoTokenizer
 
 import modelopt.torch.quantization as mtq
-from modelopt.torch.export.model_config import KV_CACHE_FP8
+from modelopt.torch.export.quant_format import KV_CACHE_FP8
 from modelopt.torch.export.quant_utils import get_quant_config
 from modelopt.torch.kernels.quantization.gemm import weight_dequant
 from modelopt.torch.quantization.nn import TensorQuantizer
