+++ modelopt/torch/export/unified_export_megatron.py
@@ -39,16 +39,6 @@ from modelopt.torch.quantization.nn.modules.tensor_quantizer import GroupedQuant
 from modelopt.torch.utils import import_plugin, warn_rank_0
 
 from .convert_hf_config import convert_hf_quant_config_format
-from .model_config import (
-    KV_CACHE_FP8,
-    KV_CACHE_NVFP4,
-    QUANTIZATION_FP8,
-    QUANTIZATION_FP8_PB_REAL,
-    QUANTIZATION_FP8_PB_WO,
-    QUANTIZATION_NONE,
-    QUANTIZATION_NVFP4,
-    QUANTIZATION_W4A16_NVFP4,
-)
 from .plugins.hf_checkpoint_utils import (
     copy_hf_ckpt_remote_code,
     copy_non_safetensor_files_from_ckpt,
@@ -65,6 +55,16 @@ from .plugins.mcore_custom import (
     save_safetensors_by_layer_index,
 )
 from .plugins.megatron_importer import GPTModelImporter, _get_mamba_conv1d
+from .quant_format import (
+    KV_CACHE_FP8,
+    KV_CACHE_NVFP4,
+    QUANTIZATION_FP8,
+    QUANTIZATION_FP8_PB_REAL,
+    QUANTIZATION_FP8_PB_WO,
+    QUANTIZATION_NONE,
+    QUANTIZATION_NVFP4,
+    QUANTIZATION_W4A16_NVFP4,
+)
 from .quant_utils import (
     get_activation_scaling_factor,
     get_kv_cache_dtype,
