+++ tests/unit/torch/export/test_offload_export.py
@@ -36,8 +36,8 @@ from _test_utils.torch.quantization.tied_modules import (
 )
 
 import modelopt.torch.quantization as mtq
-from modelopt.torch.export.model_config import KV_CACHE_FP8
 from modelopt.torch.export.model_utils import TiedWeightMap
+from modelopt.torch.export.quant_format import KV_CACHE_FP8
 from modelopt.torch.export.quant_utils import _postprocess_single_tensor
 from modelopt.torch.export.unified_export_hf import _export_quantized_weight
 from modelopt.torch.export.unified_export_hf_streaming import (
